backend/requirements.txt
#requirements needed for this project to run
#Django dependencies
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.7.0

#modelling 
torch==2.6.0
numpy==2.2.5

#API calls + environment
requests==2.32.3
python-dotenv==1.1.0

#testing 
selenium==4.32.0